SkillLens/requirements.txt
pymupdf
python-docx
endee
fastembed
langchain-google-genai
langchain-core
python-dotenv
quart
quart-cors
